move some webapp files into packages/jamtools/platforms/webapp
