+++ apps/jamtools/webapp/src/main.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import {JamToolsEngine, JamToolsProvider} from '~/core/engine/engine';
 
-import {FrontendRoutes} from './frontend_routes';
+import {FrontendRoutes} from '~/platforms/webapp/frontend_routes';
 
 type Props = {
     engine: JamToolsEngine;
